requirements.txt
aiofiles==23.1.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
blobfile==2.0.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
fastapi==0.92.0
ffmpy==0.3.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gradio==3.19.1
h11==0.14.0
html==1.13
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.11.1
idna==3.4
ipykernel==6.20.1
ipython==8.8.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.3
kiwisolver==1.4.4
linkify-it-py==2.0.0
lxml==4.9.2
markdown-it-py==2.2.0
markdownify==0.11.6
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.1
openai==0.26.1
orjson==3.8.6
outcome==1.2.0
packaging==23.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipreqs==0.4.11
platformdirs==2.6.2
plotly==5.12.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyChatGPT==0.4.3.3
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
pydantic==1.10.5
pydub==0.25.1
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.5
pytz==2022.7.1
pypiwin32==223
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
rfc3986==1.5.0
scikit-learn==1.2.0
scipy==1.10.0
selenium==4.8.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.25.0
tenacity==8.1.0
threadpoolctl==3.1.0
tiktoken==0.1.2
tokenizers==0.13.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
transformers==4.25.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2023.3
uc-micro-py==1.0.1
undetected-chromedriver==3.4.6
urllib3==1.26.13
uvicorn==0.20.0
wcwidth==0.2.5
websockets==10.4
whisper==1.1.10
wsproto==1.2.0
yarg==0.1.9
yarl==1.8.2
